--- src/backend/services/mcp/llm-client-provider.ts
@@ -1,296 +1,0 @@
-import { readFile } from "node:fs/promises";
-import type { TranscriptSegment } from "@shared/types/transcript";
-import type { LanguageModel, ModelMessage, ToolSet, TranscriptionModel } from "ai";
-import {
-  generateText,
-  Output,
-  stepCountIs,
-  streamText,
-  experimental_transcribe as transcribe,
-} from "ai";
-import { BrowserWindow } from "electron";
-import type { ZodType, z } from "zod";
-import { LLM_PROVIDER_CONFIGS } from "../../../shared/llm/llm-providers";
-import type { LLMConfig } from "../../../shared/types/llm";
-import type { HealthStatusInfo } from "../../types";
-import { formatErrorMessage } from "../../utils/error-utils";
-import { LlmStorage } from "../storage/llm-storage";
-
-type StepType =
-  | "start"
-  | "reasoning"
-  | "tool_call"
-  | "tool_result"
-  | "final_result"
-  | "tool_approval_required"
-  | "tool_denied";
-
-interface MCPStep {
-  type: StepType;
-  message?: string;
-  reasoning?: string;
-  toolName?: string;
-  serverName?: string;
-  args?: unknown;
-  result?: unknown;
-  error?: string;
-  timestamp?: number;
-}
-
-function sendStepEvent(event: MCPStep): void {
-  for (const win of BrowserWindow.getAllWindows()) {
-    if (!win.isDestroyed()) {
-      win.webContents.send("mcp:step-update", event);
-    }
-  }
-}
-
-export class LLMClientProvider {
-  public static llmClient: LLMClientProvider | null = null;
-  private static languageModel: LanguageModel;
-  private static transcriptionModel: TranscriptionModel;
-
-  private constructor() {}
-
-  static async getInstanceAsync(): Promise<LLMClientProvider | null> {
-    if (LLMClientProvider.llmClient) {
-      return LLMClientProvider.llmClient;
-    }
-
-    await LLMClientProvider.updateLanguageModelAsync();
-    await LLMClientProvider.updateTranscriptionModelAsync();
-    return LLMClientProvider.llmClient;
-  }
-
-  // Get the latest language model configuration
-  public static async updateLanguageModelAsync(): Promise<void> {
-    if (!LLMClientProvider.llmClient) {
-      LLMClientProvider.llmClient = new LLMClientProvider();
-    }
-    // retrieve LLM configuration
-    const llmConfig: LLMConfig =
-      (await LlmStorage.getInstance().getLLMConfig())?.languageModel ??
-      (() => {
-        throw new Error("[LLMClientProvider]: LLM language configuration not found");
-      })();
-
-    const config = LLM_PROVIDER_CONFIGS[llmConfig.provider];
-    if (!config || !config.defaultLanguageModel) {
-      throw new Error(`[LLMClientProvider]: Unsupported LLM provider: ${llmConfig.provider}`);
-    }
-
-    console.log(
-      `[LLMClientProvider]: updateLanguageModelAsync - configuring ${llmConfig.provider}`,
-    );
-
-    const client = config.factory({ apiKey: llmConfig.apiKey });
-    const modelName = llmConfig.model ?? config.defaultLanguageModel;
-    console.log(`[LLMClientProvider]: Using model: ${modelName}`);
-    LLMClientProvider.languageModel = client.languageModel(modelName);
-  }
-
-  public static async updateTranscriptionModelAsync(): Promise<void> {
-    if (!LLMClientProvider.llmClient) {
-      LLMClientProvider.llmClient = new LLMClientProvider();
-    }
-    // retrieve LLM configuration
-    const llmConfig: LLMConfig =
-      (await LlmStorage.getInstance().getLLMConfig())?.transcriptionModel ??
-      (() => {
-        throw new Error("[LLMClientProvider]: LLM transcription configuration not found");
-      })();
-
-    const config = LLM_PROVIDER_CONFIGS[llmConfig.provider];
-    if (!config || !config.defaultTranscriptionModel) {
-      throw new Error(`[LLMClientProvider]: Unsupported LLM provider: ${llmConfig.provider}`);
-    }
-
-    console.log(
-      `[LLMClientProvider]: updateTranscriptionModelAsync - configuring ${llmConfig.provider}`,
-    );
-
-    const client = config.factory({ apiKey: llmConfig.apiKey });
-    const modelName = llmConfig.model ?? config.defaultTranscriptionModel;
-    if (!modelName) throw new Error("[LLMClientProvider]: No transcription model specified");
-    const transcriptionModel = client.transcriptionModel?.(modelName);
-    if (!transcriptionModel)
-      throw new Error(`[LLMClientProvider]: Transcription model ${modelName} could not be created`);
-
-    LLMClientProvider.transcriptionModel = transcriptionModel;
-  }
-
-  public async generateText(messages: ModelMessage[]): Promise<string> {
-    if (!LLMClientProvider.languageModel) {
-      throw new Error("[LLMClientProvider]: LLM client not initialized");
-    }
-
-    const { text } = await generateText({
-      model: LLMClientProvider.languageModel,
-      messages: messages,
-    });
-    return text;
-  }
-
-  public async generateTextWithTools(
-    messages: ModelMessage[],
-    tools?: ToolSet,
-  ): Promise<Awaited<ReturnType<typeof generateText>>> {
-    if (!LLMClientProvider.languageModel) {
-      throw new Error("[LLMClientProvider]: LLM client not initialized");
-    }
-
-    // remove 'execute' functions from tools before passing to generateText to prevent ai sdk auto execute the tool
-    const sanitizedTools = tools
-      ? Object.fromEntries(
-          Object.entries(tools).map(([name, { execute, ...rest }]) => [name, rest]),
-        )
-      : undefined;
-
-    const response = await generateText({
-      model: LLMClientProvider.languageModel,
-      messages: messages,
-      tools: sanitizedTools,
-    });
-
-    return response;
-  }
-
-  public async generateObject<T extends ZodType>(
-    prompt: string,
-    schema: T,
-    systemPrompt?: string,
-  ): Promise<z.infer<T>> {
-    try {
-      const { output } = await generateText({
-        system: systemPrompt,
-        model: LLMClientProvider.languageModel,
-        output: Output.object({ schema }),
-        prompt: prompt,
-      });
-      return schema.parse(output);
-    } catch (error) {
-      throw new Error(
-        `[LLMClientProvider]: Failed to generate object matching schema. Prompt: "${prompt}". Schema: ${schema?.toString?.() || "[unknown schema]"}. Original error: ${error instanceof Error ? error.message : String(error)}`,
-      );
-    }
-  }
-
-  public async generateJson(prompt: string, systemPrompt?: string): Promise<string> {
-    try {
-      const { output } = await generateText({
-        system: systemPrompt,
-        model: LLMClientProvider.languageModel,
-        output: Output.json(),
-        prompt: prompt,
-      });
-      if (!output) {
-        throw new Error("No output received from LLM when generating JSON");
-      }
-      return JSON.stringify(output);
-    } catch (error) {
-      throw new Error(
-        `[LLMClientProvider]: Failed to generate object matching schema. Prompt: "${prompt}". Original error: ${error instanceof Error ? error.message : String(error)}`,
-      );
-    }
-  }
-
-  public async sendMessage(
-    message: ModelMessage[],
-    tools: ToolSet,
-  ): Promise<Awaited<ReturnType<typeof streamText>>> {
-    if (!LLMClientProvider.languageModel) {
-      throw new Error("[LLMClientProvider]: LLM client not initialized");
-    }
-
-    try {
-      return streamText({
-        model: LLMClientProvider.languageModel,
-        tools,
-        messages: message,
-        stopWhen: stepCountIs(50),
-        onFinish: (result) => sendStepEvent({ type: "final_result", message: result.finishReason }),
-        onChunk: ({ chunk }) => {
-          switch (chunk.type) {
-            case "tool-call":
-              // send an event to your orchestrator UI: "Calling tool X …"
-              sendStepEvent({ type: "tool_call", toolName: chunk.toolName, args: chunk.input });
-              break;
-            case "tool-result":
-              // show the result once the tool returns
-              sendStepEvent({
-                type: "tool_result",
-                toolName: chunk.toolName,
-                result: chunk.output,
-              });
-              break;
-          }
-        },
-      });
-    } catch (error) {
-      console.error("[LLMClientProvider]: Error in sendMessage:", error);
-      throw error;
-    }
-  }
-
-  public async transcribeAudio(filePath: string): Promise<TranscriptSegment[]> {
-    if (!LLMClientProvider.transcriptionModel) {
-      throw new Error("[LLMClientProvider]: LLM transcription client not initialized");
-    }
-    const res = await transcribe({
-      model: LLMClientProvider.transcriptionModel,
-      audio: await readFile(filePath),
-      providerOptions: {
-        openai: {
-          timestampGranularities: ["segment"],
-        },
-      },
-    });
-    if (res.segments.length === 0) {
-      return [
-        {
-          text: res.text,
-          startSecond: 0,
-          endSecond: res.durationInSeconds || 0,
-        },
-      ];
-    }
-    return res.segments.map((segment) => ({
-      text: segment.text,
-      startSecond: segment.startSecond,
-      endSecond: segment.endSecond,
-    }));
-  }
-
-  public static async checkHealthAsync(): Promise<HealthStatusInfo> {
-    // Ensure the latest model configuration
-    await LLMClientProvider.updateLanguageModelAsync();
-
-    try {
-      const providerInstance = await LLMClientProvider.getInstanceAsync();
-      if (!providerInstance) {
-        return {
-          isHealthy: false,
-          error: "LLM client provider not initialized",
-        };
-      }
-
-      const messages: ModelMessage[] = [
-        { role: "user", content: "what is your model name and version?" },
-      ];
-
-      const response = await providerInstance.generateText(messages);
-
-      console.log("[LLMClientProvider]: Health check response:", response);
-
-      return {
-        isHealthy: true,
-        successMessage: `Healthy - Model: ${response}`,
-      };
-    } catch (err) {
-      return {
-        isHealthy: false,
-        error: formatErrorMessage(err),
-      };
-    }
-  }
-}
